+++ clients/drcachesim/tools/syscall_mix.h
@@ -74,7 +74,7 @@ public:
         // Failure code counts, but only for system calls traced
         // with -record_syscall where TRACE_MARKER_TYPE_SYSCALL_FAILED
         // is recorded.
-        std::unordered_map<int, int64_t> syscall_errno_counts;
+        std::unordered_map<int64_t, int64_t> syscall_errno_counts;
     };
 
     statistics_t
